% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSI_R_package.R
\name{computeSSI}
\alias{computeSSI}
\title{A function that computes Systematicity of Search Index (SSI)}
\usage{
computeSSI(df, dfRan, participant, trial, alternative, attribute, num_alt,
  num_att, threshold, iter)
}
\arguments{
\item{df}{Object of class data frame}

\item{dfRan}{The same object used for creating random data.}

\item{participant}{Identifies each unique subject.}

\item{trial}{Identifies each unique trial.}

\item{alternative}{Represents column with eye fixations to different alternatives.}

\item{attribute}{Represents column with eye fixations to different attributes.}

\item{num_alt}{Number of alternatives in the experiment.}

\item{num_att}{Number of attributes in the experiment.}

\item{threshold}{Sets the threshold for pattern length to two or four.}

\item{iter}{Number of simulation iterations.}
}
\value{
A data set with SSI values for all identified alternative- and attribute-wise patterns.
}
\description{
Integrates all the necessary functions and computes the SSI. The number of alternatives and attributes, simulations as well as the minimum threshold for pattern length can be adjusted.
}
\examples{
#IMPORTANT! Variable names in your data set should match to the ones in the example below!

\dontrun{
dataSet <- data.frame(participant = rep(c(1:50), each = 400),
                      trial = rep(c(1:200), each = 100),
                      alternative = sample(1:4, 20000, TRUE),
                      attribute = sample(c("a","b","c","d"), 20000, TRUE))

SSI <- computeSSI(dataSet, dataSet, "participant", "trial", "alternative", "attribute", 4, 4, 4, 10)
}

}
\author{
Sonja Perkovic, \email{bnsp@leeds.ac.uk}
}
\keyword{SSI}
